SpreadsheetSelectionWidget introduced
- Base class inserted between SpreadsheetHistoryAwareStateWidget and SpreadsheetCellWidget.
- SpreadsheetViewportWidget no longer extends SpreadsheetCellWidget, but rather extends SpreadsheetSelectionWidget.
